+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.17.0 - (December 10, 2019)
+------------------
+### Added
+* Added orion-fusion-theme and generated the respective reference wdio screenshots.
+
 3.16.0 - (November 7, 2019)
 ------------------
 ### Changed